# Conflicts:
#	CarShowroomProj/src/main/java/pl/polsl/web/WorkerWebController.java
#	CarShowroomProj/src/main/resources/templates/worker.html

+++ CarShowroomProj/src/main/java/pl/polsl/web/DictionaryWebController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import pl.polsl.Data;
 import pl.polsl.ViewMode;
 import pl.polsl.controller.DictionaryController;
 import pl.polsl.model.Dictionary;
@@ -17,13 +18,26 @@ import pl.polsl.model.Dictionary;
  */
 @Controller
 public class DictionaryWebController extends BaseWebController{
-    private ViewMode viewMode;
     @Autowired
     private DictionaryController dictionaryController;
 
     @RequestMapping(value ="/dictionary")
-    public String getDictionaries(Model model){
+    public String getDictionaries(Model model) {
         model.addAttribute("dictionaries", dictionaryController.findAll());
+        if (Data.user == null) {
+            model.asMap().clear();
+            model.addAttribute("userNotLoggedIn", true);
+            return "sign_in";
+        } else if (!privilegesController.getReadPriv(Data.dictionaryModuleValue, Data.user)) {
+            model.asMap().clear();
+            model.addAttribute("forbiddenAccess", true);
+        } else {
+            model.addAttribute("dictionaries", dictionaryController.findAll());
+        }
+        analisePrivileges(Data.dictionaryModuleValue);
+        model.addAttribute("insertEnabled", insertEnabled);
+        model.addAttribute("updateEnabled", updateEnabled);
+        model.addAttribute("deleteEnabled", deleteEnabled);
         refreshMenuPrivileges(model);
         return "dictionary";
     }
@@ -34,7 +48,6 @@ public class DictionaryWebController extends BaseWebController{
         redirectAttributes.addFlashAttribute("controlsPanelVisible", true);
         redirectAttributes.addFlashAttribute("controlsDisabled", false);
         redirectAttributes.addFlashAttribute("dictionary", dictionaryController.findOne(id));
-        redirectAttributes.addFlashAttribute("dictionaries", dictionaryController.findAll());
         redirectAttributes.addFlashAttribute("types", dictionaryController.findAllTypes());
         return "redirect:/dictionary/";
     }
@@ -45,22 +58,27 @@ public class DictionaryWebController extends BaseWebController{
         redirectAttributes.addFlashAttribute("controlsPanelVisible", true);
         redirectAttributes.addFlashAttribute("controlsDisabled", true);
         redirectAttributes.addFlashAttribute("dictionary", dictionaryController.findOne(id));
-        redirectAttributes.addFlashAttribute("dictionaries", dictionaryController.findAll());
         redirectAttributes.addFlashAttribute("types", dictionaryController.findAllTypes());
         return "redirect:/dictionary/";
     }
 
     @RequestMapping(value ="/acceptModifyDictionary", method = RequestMethod.POST)
-    public String editDictionary(@RequestParam("id") int id,
+    public String addOrEditDictionary(RedirectAttributes redirectAttributes, @RequestParam("id") int id,
                                @RequestParam("type") String type,
                                @RequestParam(value = "value") String value,
-                               @RequestParam(value = "value2")String value2){
+                               @RequestParam(value = "value2")String value2,
+                                 @RequestParam(value = "value3")String value3){
         if(viewMode == ViewMode.EDIT) {
-            Dictionary dictionary = dictionaryController.updateDictionaryValue(id, type, value, value2);
+           if(dictionaryController.updateDictionaryValue(id, type, value, value2, value3) == null){
+               redirectAttributes.addFlashAttribute("databaseError");
+           }
         }
         else if(viewMode==ViewMode.INSERT) {
-            Dictionary dictionary = dictionaryController.addDictionaryValue(id, type, value, value2);
+            if(dictionaryController.addDictionaryValue(id, type, value, value2, value3)==null){
+                redirectAttributes.addFlashAttribute("databaseError");
+            }
         }
+        viewMode = ViewMode.DEFAULT;
         return "redirect:/dictionary/";
     }
 
@@ -70,14 +88,30 @@ public class DictionaryWebController extends BaseWebController{
         redirectAttributes.addFlashAttribute("controlsPanelVisible", true);
         redirectAttributes.addFlashAttribute("controlsDisabled", false);
         redirectAttributes.addFlashAttribute("dictionary", new Dictionary());
-        redirectAttributes.addFlashAttribute("dictionaries", dictionaryController.findAll());
         redirectAttributes.addFlashAttribute("types", dictionaryController.findAllTypes());
         return "redirect:/dictionary/";
     }
 
     @RequestMapping(value ="/deleteDictionary/{id}")
-    public String deleteDictionary(@PathVariable("id")int id){
-        dictionaryController.delete(id);
+    public String deleteDictionary(RedirectAttributes redirectAttributes, @PathVariable("id")int id) {
+        Dictionary dictionary = dictionaryController.findOne(id);
+        if (dictionary.getPosition() != null && dictionary.getPosition().size() != 0
+                || (dictionary.getAccessory() != null && dictionary.getAccessory().size() != 0)
+                || (dictionary.getCarName() != null && dictionary.getCarName().size() != 0)
+                || (dictionary.getCity() != null && dictionary.getCity().size() != 0)
+                || (dictionary.getCountry() != null && dictionary.getCountry().size() != 0)
+                || (dictionary.getInvoiceType() != null && dictionary.getInvoiceType().size() != 0)
+                || (dictionary.getPaymentForm() != null && dictionary.getPaymentForm().size() != 0)
+                || (dictionary.getServices() != null && dictionary.getServices().size() != 0))
+            redirectAttributes.addFlashAttribute("deleteDictionaryError", true);
+        else
+            dictionaryController.delete(id);
         return "redirect:/dictionary/";
+    }
+
+    @RequestMapping(value="/ResetDictionaryChange")
+    public  String resetChange(){
+        viewMode = ViewMode.DEFAULT;
+        return "redirect:/dictionary";
     }
 }
